chore: Rename `parsers/` to `types/`
