+++ package/android/src/main/java/com/mrousavy/camera/CameraView+RecordVideo.kt
@@ -9,9 +9,9 @@ import com.mrousavy.camera.core.MicrophonePermissionError
 import com.mrousavy.camera.core.RecorderError
 import com.mrousavy.camera.core.RecordingSession
 import com.mrousavy.camera.core.code
-import com.mrousavy.camera.parsers.Torch
-import com.mrousavy.camera.parsers.VideoCodec
-import com.mrousavy.camera.parsers.VideoFileType
+import com.mrousavy.camera.types.Torch
+import com.mrousavy.camera.types.VideoCodec
+import com.mrousavy.camera.types.VideoFileType
 import com.mrousavy.camera.utils.makeErrorMap
 import java.util.*
