+++ package/android/src/main/java/com/mrousavy/camera/CameraView+TakePhoto.kt
@@ -12,8 +12,8 @@ import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.WritableMap
 import com.mrousavy.camera.core.CameraSession
-import com.mrousavy.camera.parsers.Flash
-import com.mrousavy.camera.parsers.QualityPrioritization
+import com.mrousavy.camera.types.Flash
+import com.mrousavy.camera.types.QualityPrioritization
 import com.mrousavy.camera.utils.*
 import java.io.File
 import java.io.FileOutputStream
