+++ package/android/src/main/java/com/mrousavy/camera/core/CameraError.kt
@@ -1,7 +1,7 @@
 package com.mrousavy.camera.core
 
 import com.mrousavy.camera.core.outputs.CameraOutputs
-import com.mrousavy.camera.parsers.CameraDeviceError
+import com.mrousavy.camera.types.CameraDeviceError
 
 abstract class CameraError(
   /**
