+++ package/android/src/main/java/com/mrousavy/camera/core/CameraSession.kt
@@ -24,12 +24,12 @@ import com.mrousavy.camera.extensions.createPhotoCaptureRequest
 import com.mrousavy.camera.extensions.openCamera
 import com.mrousavy.camera.extensions.setZoom
 import com.mrousavy.camera.frameprocessor.FrameProcessor
-import com.mrousavy.camera.parsers.Flash
-import com.mrousavy.camera.parsers.Orientation
-import com.mrousavy.camera.parsers.QualityPrioritization
-import com.mrousavy.camera.parsers.VideoCodec
-import com.mrousavy.camera.parsers.VideoFileType
-import com.mrousavy.camera.parsers.VideoStabilizationMode
+import com.mrousavy.camera.types.Flash
+import com.mrousavy.camera.types.Orientation
+import com.mrousavy.camera.types.QualityPrioritization
+import com.mrousavy.camera.types.VideoCodec
+import com.mrousavy.camera.types.VideoFileType
+import com.mrousavy.camera.types.VideoStabilizationMode
 import java.io.Closeable
 import java.util.concurrent.CancellationException
 import kotlinx.coroutines.sync.Mutex
