+++ package/android/src/main/java/com/mrousavy/camera/extensions/CameraManager+openCamera.kt
@@ -8,7 +8,7 @@ import android.util.Log
 import com.mrousavy.camera.core.CameraCannotBeOpenedError
 import com.mrousavy.camera.core.CameraDisconnectedError
 import com.mrousavy.camera.core.CameraQueues
-import com.mrousavy.camera.parsers.CameraDeviceError
+import com.mrousavy.camera.types.CameraDeviceError
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 import kotlinx.coroutines.suspendCancellableCoroutine
