+++ package/android/src/main/java/com/mrousavy/camera/types/JSUnionValue.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera.parsers
+package com.mrousavy.camera.types
 
 interface JSUnionValue {
   val unionValue: String
