+++ package/android/src/main/java/com/mrousavy/camera/types/LensFacing.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera.parsers
+package com.mrousavy.camera.types
 
 import android.hardware.camera2.CameraCharacteristics
