+++ package/android/src/main/java/com/mrousavy/camera/types/PermissionStatus.kt
@@ -1,4 +1,4 @@
-package com.mrousavy.camera.parsers
+package com.mrousavy.camera.types
 
 import android.content.pm.PackageManager
